Address review concerns about potentially optimized computations.
In theory the results of the first computation could be cached and
used repeatedly. I don't think this was happening, but without
inspecting the assembly I don't really know. Using a completely
generated list of numbers in which the same memory address is not used
as input twice resolves the concern.